Ported code changes from MFT 4.17.0, Spectrum2.csv is missing from mstdump
Description:Align adabe_parser, add missing file
