+++ adb_parser/adb_db_temp.cpp
@@ -1,71 +1,3 @@
-/* 
- * Copyright (C) Jan 2019 Mellanox Technologies Ltd. All rights reserved.
- * 
- * This software is available to you under a choice of one of two
- * licenses.  You may choose to be licensed under the terms of the GNU
- * General Public License (GPL) Version 2, available from the file
- * COPYING in the main directory of this source tree, or the
- * OpenIB.org BSD license below:
- * 
- *     Redistribution and use in source and binary forms, with or
- *     without modification, are permitted provided that the following
- *     conditions are met:
- * 
- *      - Redistributions of source code must retain the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer.
- * 
- *      - Redistributions in binary form must reproduce the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer in the documentation and/or other materials
- *        provided with the distribution.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
-
- *
- * End of legal section ......................................................
- *
- *  adb_db.c - "C" wrapper module for adb_parser classes.
- *              This module provides an alternative "C" way to work with AdbInstance objects.
- *
- *  Version: $Id$
- *
- *  Author: Wasim Abu Moch (wasim@mellanox.co.il)
- */
-
-#include <vector>
-#include <stdio.h>
-#include <boost/algorithm/string.hpp>
-#include <boost/lexical_cast.hpp>
-#include "adb_db.h"
-
-#define CHECK_FIELD(field, node_w) \
-    if (((AdbInstance*)field)->isReserved()) { \
-        continue; \
-    } \
-    offset = db_field_offset(field); \
-    if (offset - node_w->node->offset < from || \
-        offset - node_w->node->offset > to) { \
-        continue; \
-    }
-
-struct node_wrapper
-{
-    AdbInstance *node;
-    vector<AdbInstance*> *fields;
-};
-
-static char err[1024] = {0};
-
-/**
- * db_create
  */
 adb_db_t* db_create()
 {
